// src/sdk/generated/wnba.d.ts
export interface components {
    schemas: {
// ...
        Game: {
            /** @description The unique ID of this game */
            GameID?: number;
            /** @description The WNBA season of the game */
            Season?: number;
            /** @description The type of season that this record corresponds to (1=Regular Season; 2=Preseason; 3=Postseason; 4=Offseason; 5=AllStar; 6=Exhibition). */
            SeasonType?: number;
            /** @description Indicates the game's status. Possible values include: Scheduled; InProgress; Final; F/OT; Suspended; Postponed; Delayed; Canceled; Forfeit; NotNecessary */
            Status?: string | null;
            /** @description The day that the game is scheduled to be played (US Eastern Time) */
            Day?: string | null;
            /** @description The date and time of the game (US Eastern Time) */
            DateTime?: string | null;
            /** @description The abbreviation [Key] of the away team */
            AwayTeam?: string | null;
            /** @description The abbreviation [Key] of the home team */
            HomeTeam?: string | null;
            /** @description The unique ID of the away team */
            AwayTeamID?: number;
            /** @description The unique ID of the home team */
            HomeTeamID?: number;
            /** @description The unique ID of the stadium */
            StadiumID?: number | null;
            /** @description Total number of points the away team scored in the game */
            AwayTeamScore?: number | null;
            /** @description Total number of points the home team scored in the game */
            HomeTeamScore?: number | null;
            /** @description The timestamp of when this record was last updated (in US Eastern time) */
            Updated?: string | null;
            /** @description The current quarter in the game. Possible values include: 1, 2, 3, 4, Half, OT, NULL */
            Quarter?: string | null;
            /** @description Number of minutes remaining in the quarter */
            TimeRemainingMinutes?: number | null;
            /** @description Number of seconds remaining in the quarter */
            TimeRemainingSeconds?: number | null;
            /** @description A globally unique ID for the game. This value is guaranteed to be unique across all sports/leagues. */
            GlobalGameID?: number;
            /** @description A globally unique ID for the away team. This value is guaranteed to be unique across all sports/leagues. */
            GlobalAwayTeamID?: number;
            /** @description A globally unique ID for the home team. This value is guaranteed to be unique across all sports/leagues. */
            GlobalHomeTeamID?: number;
            /** @description The details of the quarters (including overtime periods) for this game. */
            Quarters?: components["schemas"]["Quarter"][];
            /** @description Money line from the perspective of the home team */
            HomeTeamMoneyLine?: number | null;
            /** @description Money line from the perspective of the away team */
            AwayTeamMoneyLine?: number | null;
            /** @description The point spread (as listed for the home team) */
            PointSpread?: number | null;
            /** @description The payout for the home team beating the spread */
            PointSpreadHomeTeamPayout?: number | null;
            /** @description The payout for the away team beating the spread */
            PointSpreadAwayTeamPayout?: number | null;
            /** @description The sportsbook's total points scored over or under for the game. Note: this is also known as the Total for the game. */
            OverUnder?: number | null;
            /** @description The sportsbook's payout for the over */
            OverPayout?: number | null;
            /** @description The sportsbook's payout for the under */
            UnderPayout?: number | null;
            /** @description Indicates whether the game is over and the final score has been verified and closed out. */
            IsClosed?: boolean;
            /** @description The date and time that the game ended (US Eastern Time) */
            GameEndDateTime?: string | null;
            /** @description The GameID of the originally scheduled, postponed game, that this game was rescheduled from. This only pertains to games that are scheduled as "make up" games. */
            RescheduledFromGameID?: number | null;
            /** @description The GameID of the game that was rescheduled from this game. This only pertains to postponed games that require rescheduling. */
            RescheduledGameID?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}