// src/sdk/generated/cwbb.d.ts
export interface components {
    schemas: {
// ...
        Game: {
            /** @description The unique ID of the game */
            GameID?: number;
            /** @description The season that the game occurs in Note: season is a single year that the majority of the season occurs in */
            Season?: number;
            /** @description The type of season that this record corresponds to (1=Regular Season, 2=Preseason, 3=Postseason, 4=Offseason, 5=AllStar) */
            SeasonType?: number;
            /** @description Indicates the game's status. Possible values include: Scheduled; InProgress; Final; F/OT; Suspended; Postponed; Delayed; Canceled; Forfeit */
            Status?: string | null;
            /** @description The date the game is scheduled to occur */
            Day?: string | null;
            /** @description The timestamp of when the game is scheduled to start */
            DateTime?: string | null;
            /** @description The abbreviation [Key] of the away team */
            AwayTeam?: string | null;
            /** @description The abbreviation [Key] of the home team */
            HomeTeam?: string | null;
            /** @description The unique ID of the away team */
            AwayTeamID?: number | null;
            /** @description The unique ID of the home team */
            HomeTeamID?: number | null;
            /** @description Total number of points the away team scored in the game */
            AwayTeamScore?: number | null;
            /** @description Total number of points the home team scored in this game */
            HomeTeamScore?: number | null;
            /** @description The timestamp of when this game was last updated (in US Eastern Time) */
            Updated?: string | null;
            /** @description The current quarter of the game (Possible Values: 1, 2, 3, 4, OT, NULL) */
            Period?: string | null;
            /** @description Number of minutes remaining in the quarter */
            TimeRemainingMinutes?: number | null;
            /** @description Number of seconds remaining in the quarter */
            TimeRemainingSeconds?: number | null;
            /** @description A globally unique ID for this game. This value is guaranteed to be unique across all sports/leagues */
            GlobalGameID?: number;
            /** @description A globally unique ID for the away team. This value is guaranteed to be unique across all sports/leagues */
            GlobalAwayTeamID?: number | null;
            /** @description A globally unique ID for the home team. This value is guaranteed to be unique across all sports/leagues */
            GlobalHomeTeamID?: number | null;
            /** @description The details of the periods (quarters & overtime) for this game */
            Periods?: components["schemas"]["Period"][];
            /** @description The timestamp in UTC of when the game is scheduled to start */
            DateTimeUTC?: string | null;
            /** @description The GameID of the originally scheduled, postponed game, that this game was rescheduled from. This only pertains to games that are scheduled as "make up" games. */
            RescheduledFromGameID?: number | null;
            /** @description The GameID of the game that was rescheduled from this game. This only pertains to postponed games that require rescheduling. */
            RescheduledGameID?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}